Moved frontend utils
