+++ src/frontend/components/OptionDetailsView.tsx
@@ -4,8 +4,8 @@ import classes from "../../styles/components/OptionDetailsView.module.scss";
 import { ArrowBackIosRounded } from "@mui/icons-material";
 import { Button, IconButton } from "@mui/material";
 import { ethers } from "ethers";
-import { account, signer } from "../../utils/frontend/metamask";
-import { dismissLastToast, showToast } from "../../utils/frontend/toasting";
+import { account, signer } from "../utils/metamask";
+import { dismissLastToast, showToast } from "../utils/toasting";
 import {  OptionState, OptionWithAsset } from "../../models/option";
 import { ADDRESS0 } from "../../utils/constants";
 import { useEffect, useState } from "react";
@@ -13,8 +13,8 @@ import toast from "react-hot-toast";
 import { isExpired } from "../../datasources/options";
 import { getCachedContract } from "../../datasources/NFT/localhost";
 import { useContracts, useOptionChangingIDs } from "../../pages/_app";
-import { flavorLabels } from "../../utils/frontend/labels";
-import { getAccountDisplayValue } from "../../utils/frontend/helpers";
+import { flavorLabels } from "../utils/labels";
+import { getAccountDisplayValue } from "../utils/helpers";
 
 type OptionDetailsViewProps =
 {
