+++ src/frontend/utils/helpers.ts
