+++ src/pages/_app.tsx
@@ -3,16 +3,16 @@ import { createContext, useContext, useState, useEffect, useRef } from "react";
 import { AppProps } from "next/app";
 import Header from "../frontend/components/Header";
 import toast, { Toaster } from "react-hot-toast";
-import { connected, getSignedContract, hookMetamask, network, provider, setWindow } from "../utils/frontend/metamask";
+import { connected, getSignedContract, hookMetamask, network, provider, setWindow } from "../frontend/utils/metamask";
 import { OptionState, OptionWithAsset } from "../models/option";
 import { loadOptionWithAsset, options, setContract } from "../datasources/options";
 import { loadAllOptionsWithAsset } from "../datasources/options";
 import { NFTOpt } from "../../typechain-types";
 import NFTOptSolContract from "../../artifacts/contracts/NFTOpt.sol/NFTOpt.json";
 import addresses from "../../addresses.json";
-import { dismissLastToast, TOAST_DURATION } from "../utils/frontend/toasting";
+import { dismissLastToast, TOAST_DURATION } from "../frontend/utils/toasting";
 import { BigNumber } from "ethers";
-import { actionLabels, events } from "../utils/frontend/labels";
+import { actionLabels, events } from "../frontend/utils/labels";
 
 const OptionsHashContext       = createContext(0);
 const AccountContext           = createContext("");
